Update AdminTemplateColumns.css
I adapted this css a little to also work for within repeaters, also I have hard time understanding why float:right and display: inline-block is in there as it doesn't seem to be needed and do anything. Those are already floated.

Also wanted to mention that 70 - 29% isn't really doing anything, as PW does change value to round up to 30% anyway...

I also added a margin-left: 0px!important, as the line in the new admin theme overlapped and as there's a margin-left: -1px in the admin theme... lots of strange thing I'm not wanting to get into detail as I lost it already what is currently happening to all those wierd changed and stuff going on int he admin css.